RE-342 Split Up BetterNPC Events to its own events
